# Counter performance increase by incrementing atomics immediately when no attributes are provided