--- a/opentelemetry-sdk/src/metrics/internal/sum.rs
+++ b/opentelemetry-sdk/src/metrics/internal/sum.rs
@@ -4,6 +4,7 @@ use std::{
     sync::Mutex,
     time::SystemTime,
 };
+use std::sync::Arc;
 
 use crate::attributes::AttributeSet;
 use crate::metrics::data::{self, Aggregation, DataPoint, Temporality};
@@ -18,21 +19,15 @@ use super::{
 struct ValueMap<T: Number<T>> {
     values: Mutex<HashMap<AttributeSet, T>>,
     has_no_value_attribute_value: AtomicBool,
-    no_attribute_value: T::AtomicTracker,
-}
-
-impl<T: Number<T>> Default for ValueMap<T> {
-    fn default() -> Self {
-        ValueMap::new()
-    }
+    no_attribute_value: Arc<T::AtomicTracker>,
 }
 
 impl<T: Number<T>> ValueMap<T> {
-    fn new() -> Self {
+    fn new(no_attribute_value: Arc<T::AtomicTracker>) -> Self {
         ValueMap {
             values: Mutex::new(HashMap::new()),
             has_no_value_attribute_value: AtomicBool::new(false),
-            no_attribute_value: T::new_atomic_tracker(),
+            no_attribute_value,
         }
     }
 }
@@ -79,9 +74,9 @@ impl<T: Number<T>> Sum<T> {
     ///
     /// Each sum is scoped by attributes and the aggregation cycle the measurements
     /// were made in.
-    pub(crate) fn new(monotonic: bool) -> Self {
+    pub(crate) fn new(monotonic: bool, no_attribute_value: Arc<T::AtomicTracker>) -> Self {
         Sum {
-            value_map: ValueMap::new(),
+            value_map: ValueMap::new(no_attribute_value),
             monotonic,
             start: Mutex::new(SystemTime::now()),
         }
@@ -125,10 +120,10 @@ impl<T: Number<T>> Sum<T> {
         }
 
         let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);
-        if self
-            .value_map
-            .has_no_value_attribute_value
-            .swap(false, Ordering::AcqRel)
+        // if self
+        //     .value_map
+        //     .has_no_value_attribute_value
+        //     .swap(false, Ordering::AcqRel)
         {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
@@ -195,10 +190,10 @@ impl<T: Number<T>> Sum<T> {
 
         let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);
 
-        if self
-            .value_map
-            .has_no_value_attribute_value
-            .load(Ordering::Acquire)
+        // if self
+        //     .value_map
+        //     .has_no_value_attribute_value
+        //     .load(Ordering::Acquire)
         {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
@@ -241,7 +236,7 @@ pub(crate) struct PrecomputedSum<T: Number<T>> {
 impl<T: Number<T>> PrecomputedSum<T> {
     pub(crate) fn new(monotonic: bool) -> Self {
         PrecomputedSum {
-            value_map: ValueMap::new(),
+            value_map: ValueMap::new(Arc::new(T::new_atomic_tracker())),
             monotonic,
             start: Mutex::new(SystemTime::now()),
             reported: Mutex::new(Default::default()),
@@ -291,10 +286,10 @@ impl<T: Number<T>> PrecomputedSum<T> {
             Err(_) => return (0, None),
         };
 
-        if self
-            .value_map
-            .has_no_value_attribute_value
-            .swap(false, Ordering::AcqRel)
+        // if self
+        //     .value_map
+        //     .has_no_value_attribute_value
+        //     .swap(false, Ordering::AcqRel)
         {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
@@ -373,10 +368,10 @@ impl<T: Number<T>> PrecomputedSum<T> {
             Err(_) => return (0, None),
         };
 
-        if self
-            .value_map
-            .has_no_value_attribute_value
-            .load(Ordering::Acquire)
+        // if self
+        //     .value_map
+        //     .has_no_value_attribute_value
+        //     .load(Ordering::Acquire)
         {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),