--- a/opentelemetry-sdk/src/metrics/internal/sum.rs
+++ b/opentelemetry-sdk/src/metrics/internal/sum.rs
@@ -1,10 +1,10 @@
 use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
 use std::{
     collections::{hash_map::Entry, HashMap},
     sync::Mutex,
     time::SystemTime,
 };
-use std::sync::Arc;
 
 use crate::attributes::AttributeSet;
 use crate::metrics::data::{self, Aggregation, DataPoint, Temporality};
@@ -19,11 +19,11 @@ use super::{
 struct ValueMap<T: Number<T>> {
     values: Mutex<HashMap<AttributeSet, T>>,
     has_no_value_attribute_value: AtomicBool,
-    no_attribute_value: Arc<T::AtomicTracker>,
+    no_attribute_value: Arc<AtomicTracker<T, T::AtomicValue>>,
 }
 
 impl<T: Number<T>> ValueMap<T> {
-    fn new(no_attribute_value: Arc<T::AtomicTracker>) -> Self {
+    fn new(no_attribute_value: Arc<AtomicTracker<T, T::AtomicValue>>) -> Self {
         ValueMap {
             values: Mutex::new(HashMap::new()),
             has_no_value_attribute_value: AtomicBool::new(false),
@@ -74,7 +74,10 @@ impl<T: Number<T>> Sum<T> {
     ///
     /// Each sum is scoped by attributes and the aggregation cycle the measurements
     /// were made in.
-    pub(crate) fn new(monotonic: bool, no_attribute_value: Arc<T::AtomicTracker>) -> Self {
+    pub(crate) fn new(
+        monotonic: bool,
+        no_attribute_value: Arc<AtomicTracker<T, T::AtomicValue>>,
+    ) -> Self {
         Sum {
             value_map: ValueMap::new(no_attribute_value),
             monotonic,
@@ -120,16 +123,12 @@ impl<T: Number<T>> Sum<T> {
         }
 
         let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);
-        // if self
-        //     .value_map
-        //     .has_no_value_attribute_value
-        //     .swap(false, Ordering::AcqRel)
-        {
+        if let Some(value) = self.value_map.no_attribute_value.get_value(true) {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
                 start_time: Some(prev_start),
                 time: Some(t),
-                value: self.value_map.no_attribute_value.get_and_reset_value(),
+                value,
                 exemplars: vec![],
             });
         }
@@ -189,17 +188,12 @@ impl<T: Number<T>> Sum<T> {
         }
 
         let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);
-
-        // if self
-        //     .value_map
-        //     .has_no_value_attribute_value
-        //     .load(Ordering::Acquire)
-        {
+        if let Some(value) = self.value_map.no_attribute_value.get_value(false) {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
                 start_time: Some(prev_start),
                 time: Some(t),
-                value: self.value_map.no_attribute_value.get_value(),
+                value,
                 exemplars: vec![],
             });
         }
@@ -286,16 +280,12 @@ impl<T: Number<T>> PrecomputedSum<T> {
             Err(_) => return (0, None),
         };
 
-        // if self
-        //     .value_map
-        //     .has_no_value_attribute_value
-        //     .swap(false, Ordering::AcqRel)
-        {
+        if let Some(value) = self.value_map.no_attribute_value.get_value(true) {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
                 start_time: Some(prev_start),
                 time: Some(t),
-                value: self.value_map.no_attribute_value.get_and_reset_value(),
+                value,
                 exemplars: vec![],
             });
         }
@@ -368,16 +358,12 @@ impl<T: Number<T>> PrecomputedSum<T> {
             Err(_) => return (0, None),
         };
 
-        // if self
-        //     .value_map
-        //     .has_no_value_attribute_value
-        //     .load(Ordering::Acquire)
-        {
+        if let Some(value) = self.value_map.no_attribute_value.get_value(false) {
             s_data.data_points.push(DataPoint {
                 attributes: AttributeSet::default(),
                 start_time: Some(prev_start),
                 time: Some(t),
-                value: self.value_map.no_attribute_value.get_value(),
+                value,
                 exemplars: vec![],
             });
         }