Fixes #
Design discussion issue (if applicable) #

## Changes

Previously, atomics were managed by the `ValueMap` itself in the `measure()` function. This meant that every `counter.add(1, &[])` call would have to go through several boxed function calls, plus the non-zero cost of creating an empty attribute set.  Moving the incrementing of atomics in the no attribute into the SDK's counter implementation directly bypasses all of these costs for significantly faster counter addition.

Since logic for managing these atomics was moved into multiple call sites, the original `AtomicTracker` trait was renamed to `AtomicValue` to better represent what the trait is calling.  A new `AtomicTracker` struct was created to manage both the `AtomicValue` and an `AtomicBool` so that aggregations know if they have a no-attribute value that should be exported or not.

## Benchmarks

main:
```
Counter/AddNoAttrs      time:   [29.452 ns 29.506 ns 29.559 ns]
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

Counter/AddNoAttrsDelta time:   [29.383 ns 29.513 ns 29.678 ns]
Found 12 outliers among 100 measurements (12.00%)
  6 (6.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe

Counter/AddOneAttr      time:   [139.60 ns 140.23 ns 141.07 ns]
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe

Counter/AddOneAttrDelta time:   [142.81 ns 143.27 ns 143.87 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
```

This PR:
```
Counter/AddNoAttrs      time:   [6.9870 ns 7.0010 ns 7.0183 ns]
                        change: [-76.475% -76.331% -76.213%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
Counter/AddNoAttrsDelta time:   [6.9716 ns 6.9944 ns 7.0209 ns]
                        change: [-76.366% -76.236% -76.096%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  5 (5.00%) high mild
  4 (4.00%) high severe
Counter/AddOneAttr      time:   [139.78 ns 140.11 ns 140.45 ns]
                        change: [-0.2692% +0.4519% +1.3838%] (p = 0.28 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
  4 (4.00%) high severe
Counter/AddOneAttrDelta time:   [140.96 ns 141.41 ns 141.85 ns]
                        change: [-2.1149% -1.5107% -0.9781%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 16 outliers among 100 measurements (16.00%)
  4 (4.00%) low severe
  4 (4.00%) low mild
  8 (8.00%) high mild
```

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [X] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)