@cijothomas 2024-02-21
> "@KallDrexx, nit - do you think including the `no-attribute scenario` text in the title would provide more clarity on what this PR is targeting?

Will also include some recommendations in the docs. (I'll make them soon)