@cijothomas 2024-05-11
The change itself is good, and nicely optimized the 0 attribute case to be as fast as possible. I'll not be merging this PR as-is, as I am doing some refactor on the Metrics SDK.

Discussed with original author @KallDrexx - as part of refactoring, the ideas from this PR will be folded into that PR. Many thanks for working on this. This is a good solution for high perf needs until we get the full notion of bound instruments.